// layer/authoring.generated.ts
export const componentImplementationByName = {
  MdcAside: {
    componentName: "MdcAside",
    props: {
      label: {
        required: false,
        types: ["string"],
      },
      appearance: {
        required: false,
        types: ["string"],
        options: ["quiet", "tint"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcError: {
    componentName: "MdcError",
    props: {
      title: {
        required: false,
        types: ["string"],
      },
      icon: {
        required: false,
        types: ["string"],
      },
      appearance: {
        required: false,
        types: ["string"],
        options: ["quiet", "tint"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcExcerpt: {
    componentName: "MdcExcerpt",
    props: {
      label: {
        required: false,
        types: ["string"],
      },
      source: {
        required: false,
        types: ["string"],
      },
      appearance: {
        required: false,
        types: ["string"],
        options: ["quiet", "tint"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcIdea: {
    componentName: "MdcIdea",
    props: {
      title: {
        required: false,
        types: ["string"],
      },
      icon: {
        required: false,
        types: ["string"],
      },
      appearance: {
        required: false,
        types: ["string"],
        options: ["quiet", "tint"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcNote: {
    componentName: "MdcNote",
    props: {
      title: {
        required: false,
        types: ["string"],
      },
      icon: {
        required: false,
        types: ["string"],
      },
      appearance: {
        required: false,
        types: ["string"],
        options: ["quiet", "tint"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcSuccess: {
    componentName: "MdcSuccess",
    props: {
      title: {
        required: false,
        types: ["string"],
      },
      icon: {
        required: false,
        types: ["string"],
      },
      appearance: {
        required: false,
        types: ["string"],
        options: ["quiet", "tint"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcWarning: {
    componentName: "MdcWarning",
    props: {
      title: {
        required: false,
        types: ["string"],
      },
      icon: {
        required: false,
        types: ["string"],
      },
      appearance: {
        required: false,
        types: ["string"],
        options: ["quiet", "tint"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcColumn: {
    componentName: "MdcColumn",
    props: {
      size: {
        required: false,
        types: ["string", "object"],
        default: "md",
        options: ["sm", "md", "lg"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcInfo: {
    componentName: "MdcInfo",
    props: {
      title: {
        required: false,
        types: ["string"],
      },
      icon: {
        required: false,
        types: ["string"],
      },
      appearance: {
        required: false,
        types: ["string"],
        options: ["quiet", "tint"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
  MdcLayout: {
    componentName: "MdcLayout",
    props: {
      type: {
        required: false,
        types: ["string", "object"],
        default: "default",
        options: ["default", "card", "border", "border-dashed", "outline", "outline-dashed"],
      },
      class: {
        required: false,
        types: ["complex"],
      },
    },
    slots: ["default"],
  },
} as const;